requirements.txt
fastapi>=0.95.9
google-genai>=0.1.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.0.0
webdriver-manager>=4.0.0
requests>=2.31.0
supabase>=2.3.0
py-pdf-parser==0.13.0
